Adds support for kubernetes node pool taints
