+++ digitalocean/resource_digitalocean_kubernetes_node_pool.go
@@ -4,13 +4,11 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/digitalocean/godo"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
 // to distinguish between a node pool resource and the default pool from the cluster
@@ -29,151 +27,7 @@ func resourceDigitalOceanKubernetesNodePool() *schema.Resource {
 		},
 		SchemaVersion: 1,
 
-		Schema: nodePoolResourceSchema(),
-	}
-}
-
-func nodePoolResourceSchema() map[string]*schema.Schema {
-	s := nodePoolSchema()
-
-	// add the cluster id
-	s["cluster_id"] = &schema.Schema{
-		Type:         schema.TypeString,
-		Required:     true,
-		ValidateFunc: validation.NoZeroValues,
-		ForceNew:     true,
-	}
-
-	// remove the id when this is used in a specific resource
-	// not as a child
-	delete(s, "id")
-
-	return s
-}
-
-func nodePoolSchema() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"id": {
-			Type:     schema.TypeString,
-			Computed: true,
-		},
-		"name": {
-			Type:         schema.TypeString,
-			Required:     true,
-			ValidateFunc: validation.NoZeroValues,
-		},
-
-		"size": {
-			Type:         schema.TypeString,
-			Required:     true,
-			ForceNew:     true,
-			ValidateFunc: validation.NoZeroValues,
-		},
-
-		"actual_node_count": {
-			Type:     schema.TypeInt,
-			Computed: true,
-		},
-
-		"node_count": {
-			Type:         schema.TypeInt,
-			Optional:     true,
-			ValidateFunc: validation.IntAtLeast(1),
-			DiffSuppressFunc: func(key, old, new string, d *schema.ResourceData) bool {
-				nodeCountKey := "node_count"
-				actualNodeCountKey := "actual_node_count"
-
-				// Since this schema is shared between the node pool resource
-				// and as the node pool sub-element of the cluster resource,
-				// we need to check for both variants of the incoming key.
-				keyParts := strings.Split(key, ".")
-				if keyParts[0] == "node_pool" {
-					npKeyParts := keyParts[:len(keyParts)-1]
-					nodeCountKeyParts := append(npKeyParts, "node_count")
-					nodeCountKey = strings.Join(nodeCountKeyParts, ".")
-					actualNodeCountKeyParts := append(npKeyParts, "actual_node_count")
-					actualNodeCountKey = strings.Join(actualNodeCountKeyParts, ".")
-				}
-
-				// If node_count equals actual_node_count already, then
-				// suppress the diff.
-				if d.Get(nodeCountKey).(int) == d.Get(actualNodeCountKey).(int) {
-					return true
-				}
-
-				// Otherwise suppress the diff only if old equals new.
-				return old == new
-			},
-		},
-
-		"auto_scale": {
-			Type:     schema.TypeBool,
-			Optional: true,
-			Default:  false,
-		},
-
-		"min_nodes": {
-			Type:         schema.TypeInt,
-			Optional:     true,
-			ValidateFunc: validation.IntAtLeast(1),
-		},
-
-		"max_nodes": {
-			Type:     schema.TypeInt,
-			Optional: true,
-		},
-
-		"tags": tagsSchema(),
-
-		"labels": {
-			Type:     schema.TypeMap,
-			Optional: true,
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-
-		"nodes": nodeSchema(),
-	}
-}
-
-func nodeSchema() *schema.Schema {
-	return &schema.Schema{
-		Type:     schema.TypeList,
-		Computed: true,
-		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{
-				"id": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-
-				"name": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-
-				"status": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-
-				"droplet_id": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-
-				"created_at": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-
-				"updated_at": {
-					Type:     schema.TypeString,
-					Computed: true,
-				},
-			},
-		},
+		Schema: nodePoolSchema(true),
 	}
 }
 
@@ -189,6 +43,7 @@ func resourceDigitalOceanKubernetesNodePoolCreate(ctx context.Context, d *schema
 		"auto_scale": d.Get("auto_scale"),
 		"min_nodes":  d.Get("min_nodes"),
 		"max_nodes":  d.Get("max_nodes"),
+		"taint":      d.Get("taint"),
 	}
 
 	pool, err := digitaloceanKubernetesNodePoolCreate(client, rawPool, d.Get("cluster_id").(string))
@@ -224,6 +79,7 @@ func resourceDigitalOceanKubernetesNodePoolRead(ctx context.Context, d *schema.R
 	d.Set("min_nodes", pool.MinNodes)
 	d.Set("max_nodes", pool.MaxNodes)
 	d.Set("nodes", flattenNodes(pool.Nodes))
+	d.Set("taint", flattenNodePoolTaints(pool.Taints))
 
 	// Assign a node_count only if it's been set explicitly, since it's
 	// optional and we don't want to update with a 0 if it's not set.
@@ -341,6 +197,7 @@ func digitaloceanKubernetesNodePoolCreate(client *godo.Client, pool map[string]i
 		AutoScale: pool["auto_scale"].(bool),
 		MinNodes:  pool["min_nodes"].(int),
 		MaxNodes:  pool["max_nodes"].(int),
+		Taints:    expandNodePoolTaints(pool["taint"].([]interface{})),
 	}
 
 	p, _, err := client.Kubernetes.CreateNodePool(context.Background(), clusterID, req)
@@ -385,6 +242,11 @@ func digitaloceanKubernetesNodePoolUpdate(client *godo.Client, pool map[string]i
 
 	if pool["labels"] != nil {
 		req.Labels = expandLabels(pool["labels"].(map[string]interface{}))
+	}
+
+	if pool["taint"] != nil {
+		t := expandNodePoolTaints(pool["taint"].([]interface{}))
+		req.Taints = &t
 	}
 
 	p, resp, err := client.Kubernetes.UpdateNodePool(context.Background(), clusterID, poolID, req)
@@ -483,84 +345,4 @@ func waitForKubernetesNodePoolDelete(client *godo.Client, id string, poolID stri
 	}
 
 	return fmt.Errorf("Timeout waiting to delete nodepool")
-}
-
-func expandLabels(labels map[string]interface{}) map[string]string {
-	expandedLabels := make(map[string]string)
-	if labels != nil {
-		for key, value := range labels {
-			expandedLabels[key] = value.(string)
-		}
-	}
-	return expandedLabels
-}
-
-func flattenLabels(labels map[string]string) map[string]interface{} {
-	flattenedLabels := make(map[string]interface{})
-	if labels != nil {
-		for key, value := range labels {
-			flattenedLabels[key] = value
-		}
-	}
-	return flattenedLabels
-}
-
-func expandNodePools(nodePools []interface{}) []*godo.KubernetesNodePool {
-	expandedNodePools := make([]*godo.KubernetesNodePool, 0, len(nodePools))
-	for _, rawPool := range nodePools {
-		pool := rawPool.(map[string]interface{})
-		cr := &godo.KubernetesNodePool{
-			ID:        pool["id"].(string),
-			Name:      pool["name"].(string),
-			Size:      pool["size"].(string),
-			Count:     pool["node_count"].(int),
-			AutoScale: pool["auto_scale"].(bool),
-			MinNodes:  pool["min_nodes"].(int),
-			MaxNodes:  pool["max_nodes"].(int),
-			Tags:      expandTags(pool["tags"].(*schema.Set).List()),
-			Labels:    expandLabels(pool["labels"].(map[string]interface{})),
-			Nodes:     expandNodes(pool["nodes"].([]interface{})),
-		}
-
-		expandedNodePools = append(expandedNodePools, cr)
-	}
-
-	return expandedNodePools
-}
-
-func expandNodes(nodes []interface{}) []*godo.KubernetesNode {
-	expandedNodes := make([]*godo.KubernetesNode, 0, len(nodes))
-	for _, rawNode := range nodes {
-		node := rawNode.(map[string]interface{})
-		n := &godo.KubernetesNode{
-			ID:   node["id"].(string),
-			Name: node["name"].(string),
-		}
-
-		expandedNodes = append(expandedNodes, n)
-	}
-
-	return expandedNodes
-}
-
-func flattenNodes(nodes []*godo.KubernetesNode) []interface{} {
-	flattenedNodes := make([]interface{}, 0)
-	if nodes == nil {
-		return flattenedNodes
-	}
-
-	for _, node := range nodes {
-		rawNode := map[string]interface{}{
-			"id":         node.ID,
-			"name":       node.Name,
-			"status":     node.Status.State,
-			"droplet_id": node.DropletID,
-			"created_at": node.CreatedAt.UTC().String(),
-			"updated_at": node.UpdatedAt.UTC().String(),
-		}
-
-		flattenedNodes = append(flattenedNodes, rawNode)
-	}
-
-	return flattenedNodes
 }
